+++ src/layouts/MainAppLayout/Header/AccountAvatar/styles.js
@@ -1,26 +1,25 @@
-export const avatarProps = {
-  sx: {
-    width: {
-      mobileSmall: "28px",
-      mobile: "36px",
-      tablet: "36px",
-      laptop: "36px",
-      desktop: "36px",
-      desktopLarge: "45px",
-    },
-    height: {
-      mobileSmall: "28px",
-      mobile: "36px",
-      tablet: "36px",
-      laptop: "36px",
-      desktop: "36px",
-      desktopLarge: "45px",
-    },
+const styles = {
+  avatarProps: {
+    sx: {
+      width: {
+        mobileSmall: "28px",
+        mobile: "36px",
+        tablet: "36px",
+        laptop: "36px",
+        desktop: "36px",
+        desktopLarge: "45px",
+      },
+      height: {
+        mobileSmall: "28px",
+        mobile: "36px",
+        tablet: "36px",
+        laptop: "36px",
+        desktop: "36px",
+        desktopLarge: "45px",
+      },
+    }
   },
-};
-
-export const gridProps = {
-    item: true,
+  gridProps: {
     sx: {
       position: "absolute",
       bottom: 0,
@@ -44,4 +43,55 @@ export const gridProps = {
         desktopLarge: 11.25,
       },
     }
-  };
+  }
+}
+
+// export const avatarProps = {
+//   sx: {
+//     width: {
+//       mobileSmall: "28px",
+//       mobile: "36px",
+//       tablet: "36px",
+//       laptop: "36px",
+//       desktop: "36px",
+//       desktopLarge: "45px",
+//     },
+//     height: {
+//       mobileSmall: "28px",
+//       mobile: "36px",
+//       tablet: "36px",
+//       laptop: "36px",
+//       desktop: "36px",
+//       desktopLarge: "45px",
+//     },
+//   },
+// };
+
+// export const gridProps = {
+//     item: true,
+//     sx: {
+//       position: "absolute",
+//       bottom: 0,
+//       right: 0,
+//       bgcolor: "#12D18E",
+//       borderRadius: "50%",
+//       width: {
+//         mobileSmall: 7,
+//         mobile: 9,
+//         tablet: 9,
+//         laptop: 9,
+//         desktop: 9,
+//         desktopLarge: 11.25,
+//       },
+//       height: {
+//         mobileSmall: 7,
+//         mobile: 9,
+//         tablet: 9,
+//         laptop: 9,
+//         desktop: 9,
+//         desktopLarge: 11.25,
+//       },
+//     }
+//   };
+
+export default styles;
